refactor(sidebar): extract AvatarBox, InfoContent, InfoMoreContent, and SideBarInfo from SideBar (#53)
