+++ src/components/side-bar/index.tsx
@@ -1,3 +1,31 @@
-import { SideBar } from "./side-bar";
+"use client";
+
+import React, { useState, useRef } from 'react';
+import SideBarInfo from './side-bar-info';
+import SocialList from './social-list';
+import ContactsList from './contact-list';
+
+export const SideBar: React.FC = () => {
+  const [isActive, setIsActive] = useState(false);
+  const sideBarRef = useRef<HTMLDivElement>(null);
+
+  const handleSidebarToggle = () => {
+    setIsActive(prevState => !prevState);
+  };
+  
+  const sideBarState = `sidebar ${isActive ? 'active' : ''}`;
+
+  return (
+    <aside className={sideBarState} ref={sideBarRef} data-sidebar>
+      <SideBarInfo onToggle={handleSidebarToggle} />
+      <div className="sidebar-info-more">
+        <div className="separator"></div>
+        <ContactsList />
+        <div className="separator"></div>
+        <SocialList />
+      </div>
+    </aside>
+  );
+}
 
 export default SideBar;
