+++ src/components/side-bar/info-content.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+const InfoContent: React.FC = () => (
+  <div className="info-content">
+    <h1 className="name" title="Hugo ChunHo Lin">Hugo ChunHo Lin</h1>
+    <p className="title">
+      <strong>Day ONE ⚡️</strong>
+    </p>
+  </div>
+);
+
+export default InfoContent;
